refactor: extract utils; prettify
